cilium-dbg,bgp: Move BGP printers to common BGP package
So that the printers can be used outside of the cilium-dbg
command package as well.
