+++ cilium-dbg/cmd/bgp_peer_get.go
@@ -7,15 +7,12 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"sort"
-	"strings"
 	"text/tabwriter"
-	"time"
 
 	"github.com/spf13/cobra"
 
 	"github.com/cilium/cilium/api/v1/client/bgp"
-	"github.com/cilium/cilium/api/v1/models"
+	"github.com/cilium/cilium/pkg/bgpv1/api"
 	"github.com/cilium/cilium/pkg/command"
 )
 
@@ -40,43 +37,10 @@ var BgpPeersCmd = &cobra.Command{
 				Fatalf("error getting output in JSON: %s\n", err)
 			}
 		} else {
-			printSummary(res.GetPayload())
+			w := NewTabWriter()
+			api.PrintBGPPeersTable(w, res.GetPayload(), true)
 		}
 	},
-}
-
-func printSummary(peers []*models.BgpPeer) {
-	// get new tab writer with predefined defaults
-	w := NewTabWriter()
-
-	// sort by local AS, if peers from same AS then sort by peer address.
-	sort.Slice(peers, func(i, j int) bool {
-		return peers[i].LocalAsn < peers[j].LocalAsn || peers[i].PeerAddress < peers[j].PeerAddress
-	})
-
-	fmt.Fprintln(w, "Local AS\tPeer AS\tPeer Address\tSession\tUptime\tFamily\tReceived\tAdvertised")
-	for _, peer := range peers {
-		fmt.Fprintf(w, "%d\t", peer.LocalAsn)
-		fmt.Fprintf(w, "%d\t", peer.PeerAsn)
-		fmt.Fprintf(w, "%s:%d\t", peer.PeerAddress, peer.PeerPort)
-		fmt.Fprintf(w, "%s\t", peer.SessionState)
-
-		// Time is rounded to nearest second
-		fmt.Fprintf(w, "%s\t", time.Duration(peer.UptimeNanoseconds).Round(time.Second).String())
-
-		for i, afisafi := range peer.Families {
-			if i > 0 {
-				// move by 5 tabs to align with afi-safi
-				fmt.Fprint(w, strings.Repeat("\t", 5))
-			}
-			// AFI and SAFI are concatenated for brevity
-			fmt.Fprintf(w, "%s/%s\t", afisafi.Afi, afisafi.Safi)
-			fmt.Fprintf(w, "%d\t", afisafi.Received)
-			fmt.Fprintf(w, "%d\t", afisafi.Advertised)
-			fmt.Fprintf(w, "\n")
-		}
-	}
-	w.Flush()
 }
 
 // NewTabWriter initialises tabwriter.Writer with following defaults
