+++ cilium-dbg/cmd/bgp_route_policy_get.go
@@ -6,14 +6,12 @@ package cmd
 import (
 	"errors"
 	"fmt"
-	"sort"
-	"strings"
 
 	"github.com/spf13/cobra"
 	"k8s.io/utils/ptr"
 
 	"github.com/cilium/cilium/api/v1/client/bgp"
-	"github.com/cilium/cilium/api/v1/models"
+	"github.com/cilium/cilium/pkg/bgpv1/api"
 	"github.com/cilium/cilium/pkg/command"
 )
 
@@ -51,83 +49,10 @@ var BgpRoutePoliciesCmd = &cobra.Command{
 				Fatalf("error getting output in JSON: %s\n", err)
 			}
 		} else {
-			printBGPRoutePoliciesTable(res.GetPayload())
+			w := NewTabWriter()
+			api.PrintBGPRoutePoliciesTable(w, res.GetPayload())
 		}
 	},
-}
-
-func printBGPRoutePoliciesTable(policies []*models.BgpRoutePolicy) {
-	// get new tab writer with predefined defaults
-	w := NewTabWriter()
-
-	// sort by router ASN, if policies from same ASN then sort by policy name.
-	sort.Slice(policies, func(i, j int) bool {
-		return policies[i].RouterAsn < policies[j].RouterAsn && policies[i].Name < policies[j].Name
-	})
-
-	fmt.Fprintln(w, "VRouter\tPolicy Name\tType\tMatch Peers\tMatch Families\tMatch Prefixes (Min..Max Len)\tRIB Action\tPath Actions")
-	for _, policy := range policies {
-		fmt.Fprintf(w, "%d\t", policy.RouterAsn)
-		fmt.Fprintf(w, "%s\t", policy.Name)
-		fmt.Fprintf(w, "%s\t", policy.Type)
-
-		for i, stmt := range policy.Statements {
-			if i > 0 {
-				fmt.Fprint(w, strings.Repeat("\t", 3))
-			}
-			fmt.Fprintf(w, "%s\t", formatStringArray(stmt.MatchNeighbors))
-			fmt.Fprintf(w, "%s\t", formatStringArray(formatFamilies(stmt.MatchFamilies)))
-			fmt.Fprintf(w, "%s\t", formatStringArray(formatMatchPrefixes(stmt.MatchPrefixes)))
-			fmt.Fprintf(w, "%s\t", stmt.RouteAction)
-			fmt.Fprintf(w, "%s\t", formatStringArray(formatPathActions(stmt)))
-			fmt.Fprintf(w, "\n")
-		}
-		if len(policy.Statements) == 0 {
-			fmt.Fprintf(w, "\n")
-		}
-	}
-	w.Flush()
-}
-
-func formatStringArray(arr []string) string {
-	if len(arr) == 1 {
-		return arr[0]
-	}
-	res := ""
-	for _, str := range arr {
-		res += "{" + str + "} "
-	}
-	return strings.TrimSpace(res)
-}
-
-func formatFamilies(families []*models.BgpFamily) []string {
-	var res []string
-	for _, f := range families {
-		res = append(res, fmt.Sprintf("%s/%s", f.Afi, f.Safi))
-	}
-	return res
-}
-
-func formatMatchPrefixes(pfxs []*models.BgpRoutePolicyPrefixMatch) []string {
-	var res []string
-	for _, p := range pfxs {
-		res = append(res, fmt.Sprintf("%s (%d..%d)", p.Cidr, p.PrefixLenMin, p.PrefixLenMax))
-	}
-	return res
-}
-
-func formatPathActions(stmt *models.BgpRoutePolicyStatement) []string {
-	var res []string
-	if stmt.SetLocalPreference >= 0 {
-		res = append(res, fmt.Sprintf("SetLocalPreference: %d", stmt.SetLocalPreference))
-	}
-	if len(stmt.AddCommunities) > 0 {
-		res = append(res, fmt.Sprintf("AddCommunities: %v", stmt.AddCommunities))
-	}
-	if len(stmt.AddLargeCommunities) > 0 {
-		res = append(res, fmt.Sprintf("AddLargeCommunities: %v", stmt.AddLargeCommunities))
-	}
-	return res
 }
 
 func init() {
